Made it working
Now the goal is to make it better. Currently, to make the edges, it is
checking if the 0 is next to a 1, and making the 1 an edge. However, I
think that if we check to see if the 1 is next to more than 1 edge, it
would eliminate the extraneous points.