It now works
It can be quite slow, so I have narrowed it down. The slow part occurs
where the array is “filled,” as in getting its holes filled in. The
rest is quite fast: 1.3 seconds at 2000x2000.
The goals is now to find a way to fill in the board much faster.